Calculate percentage change in value
* In order to determine whether the company should have been invested in
(e.g. was it a buy or sell) we need to see how it's value changes
compared with the sp500.
* For this we take the starting price (initialized as False and then set
on the first iteration through the data and that value is used as the
base). Then we append to our dataframe the percent change from that
value at every date.

https://www.youtube.com/watch?annotation_id=annotation_1138849257&feature=iv&index=8&list=PLQVvvaa0QuDd0flgGphKCej-9jp-QdzZ3&src_vid=PMAwBh0nrds&v=QJKJBVUywDM